+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Upcoming
+# 0.8.0
 
 ## Breaking changes
 * Moved `input_text` capability from the text edit example into `bevy_immediate_ui` as `text_input::CapabilityUiTextInput` and `ImmUiTextInput`